Handle share intent in ProfileManagerActivity
